활동 당일인 모임 방장, 참여자들에게 알림 생성 기능

+++ src/main/java/com/post_show_blues/vine/dto/meeting/MeetingDTO.java
@@ -47,7 +47,7 @@ public class MeetingDTO {
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy/MM/dd HH:mm", timezone = "Asia/Seoul")
     private LocalDateTime reqDeadline;
 
-    private int dDay;
+    private Long dDay;
 
     private String chatLink;
